openpilot v0.10.1-piV release-pi
date: 2025-09-04T19:57:21
master commit: a6c999d343b73f2331febbb47afe29c694ca3c10

+++ cereal/messaging/socketmaster.cc
@@ -33,7 +33,7 @@ MessageContext message_context;
 struct SubMaster::SubMessage {
   std::string name;
   SubSocket *socket = nullptr;
-  float freq = 0.0f;
+  int freq = 0;
   bool updated = false, alive = false, valid = false, ignore_alive;
   uint64_t rcv_time = 0, rcv_frame = 0;
   void *allocated_msg_reader = nullptr;
